migrate delete dialog in the members-list
